Various fixes and improvements - fix for S3M arpeggio getting dorked up after instrument change. - fix for song loop detection not working as expected - introduced new playback context-based sytem - integrated against updated gosound library - better error reporting

+++ internal/format/s3m/playback/effect/effect_arpeggio.go
@@ -25,8 +25,8 @@ func (e Arpeggio) Start(cs intf.Channel, ss intf.Song) {
 func (e Arpeggio) Tick(cs intf.Channel, ss intf.Song, currentTick int) {
 	mem := cs.GetMemory().(*channel.Memory)
 	xy := mem.LastNonZero(uint8(e))
-	x := (xy >> 4) - 8
-	y := (xy & 0x0f) - 8
+	x := int8(xy>>4) - 8
+	y := int8(xy&0x0f) - 8
 	doArpeggio(cs, currentTick, x, y)
 }
